% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_data.R
\docType{data}
\name{ADL}
\alias{ADL}
\title{Activities of Daily Living - dichotomous example data}
\format{
Activities of Daily Living (V06-V39: 1= yes; 0= no):
\itemize{
  \item V06: Transfer from bed independently.
  \item V07: Stand up from a chair/wheelchair independently.
  \item V08: Walk independently.
  \item V09: Stand independently.
  \item V11: micro-change positions in bed independently.
  \item V12: completely change position in bed independently.
  \item V22: Wash at a washbasin independently (without back and feet).
  \item V23: Wash back and feet independently
  \item V24: Take a shower independently
  \item V25: dress and undress the upper body independently
  \item V26: put on shoes and stockings independently
  \item V27: take a meal independently
  \item V28: drink independently from a prepared cup
  \item V36: use the toilet independently
  \item V39: intimate hygiene independently
  \item age: age (1= <65; 2= 65-74; 3= 75-84; 4= 85-94; 5= >=95)
  \item sex: age (1= female; 2= male)}
}
\source{
Data from Grebe (2013)
}
\usage{
ADL
}
\description{
ADL date from N=591 nursing home residents from Germany
}
\references{
Grebe C (2013). Pflegeaufwand und Personalbemessung in der
 stationären Langzeitpflege: Entwicklung eines empirischen
  Fallgruppensystems auf der Basis von Bewohnercharakteristika.
   3-Länder-Konferenz Pflege & Pflegewissenschaft, Konstanz.
}
\keyword{datasets}
