% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pattern_rules.R
\name{apply_pattern_rules}
\alias{apply_pattern_rules}
\title{selects item patterns based on defined rules for item combinations}
\usage{
apply_pattern_rules(
  pattern_length = NULL,
  full = NULL,
  forced_items = NULL,
  rules = NULL
)
}
\arguments{
\item{pattern_length}{a numeric vector with the alloewd lengths of the selected patterns (scale lengths)}

\item{full}{a numeric vector containing the the source for the pattern combinations, typically the indices of the items in the referring dataset}

\item{forced_items}{a numeric vector of items that are forced to occur in every selected pattern}

\item{rules}{a list of type combination_rules defining rules for pattern selection}
}
\value{
a list of numeric vectors containing the selected item patterns that match the defined rules of forced_items and/or rules.
}
\description{
selects item patterns based on defined rules for item combinations
}
\examples{
data(ADL)
forced <- c(6,22)
rules_object <- list() # rules-Object
rules_object[[1]] <- list("min", 1, 8:9)
rules_object[[1]] <- list("min", 1, 14:15)
rules_object[[1]] <- list("max", 3, 1:6)
rules_object[[7]] <- list("forbidden", c(8,9))
final_patterns <- apply_pattern_rules(pattern_length = 5:7,
  full=1:length(ADL), forced_items = forced, rules= rules_object)
}
