% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_combo_rules.R
\name{check_combo_rules}
\alias{check_combo_rules}
\title{selects item patterns based on defined rules for item combinations}
\usage{
check_combo_rules(full, forced_items = NULL, rules = NULL)
}
\arguments{
\item{full}{a list of numeric vectors containing the full set of patterns to be analysed}

\item{forced_items}{a numeric vector of items that are forced to occur in every selected pattern}

\item{rules}{a list of type combination_rules defining rules for pattern selection}
}
\value{
a list of numeric vectors containing the selected item patterns that match the defined rules of forced_items and/or rules.
}
\description{
selects item patterns based on defined rules for item combinations
}
