% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_data.R
\docType{data}
\name{cognition}
\alias{cognition}
\title{cognition - polytomous example data
data measured with the FACT-cog (subscale perceived cognitive functioning)
sample size: N=1009}
\format{
Items 1-20: FACT-cog Items. Questions starting with
 In the past 7 days...
 Answer categories: 0= never; 1= about once a week; 2= two to three times
 a week;  3=nearly every day; 4= several times a day
\itemize{
  \item A1_CogA1: I have had trouble forming thoughts
  \item A3_CogA3: My thinking has been slow
  \item C7_PC8: I have had trouble concentrating
  \item M9_PC10: I have had trouble finding my way to a familiar place
  \item M10_PC11: I have had trouble remembering where I put things, like my keys or my wallet
  \item M12_NQCOG68: I have had trouble remembering new information, like phone numbers or simple instructions
  \item V13_PC14: I have had trouble recalling the name of an object while talking to
  \item V15_PC16: I have had trouble finding the right word(s) to express
  \item V16_CogV16: I have used the wrong word when I referred to an object
  \item V17_CogV17b: I have had trouble saying what I mean in conversations with others
  \item F19_NQCOG69: I have walked into a room and forgot what I meant to get or do there
  \item F23_NQCOG77: I had to work really hard to pay attention or I would make a mistake.
  \item F24_PC26: I have forgotten names of people soon after being introduced
  \item F25_PC28: My reactions in everyday situations have been
  \item C31_PC36: I have had to work harder than usual to keep track of what I was
  \item C32_PC37: My thinking has been slower than usual
  \item C33a_PC38: I have had to work harder than usual to express myself clearly
  \item C33c_PC40: I have had to use written lists more often than usual so I would not forget things
  \item MT1_PC41:  I have had trouble keeping track of what I was doing when interrupted
  \item MT2_PC42: I have had trouble shifting back and forth between different activities that require thinking
  \item CogPCI: Subscale Score: Percieved Cognitive Impairments (CogPCI), Range 0-72
  \item SD1: What is your age?
  \item SD2: What is your gender? (1= male; 2= female)
  \item SD3: Are you of Hispanic, Latino, or Spanish origin? (0= no; 1= yes)
  \item SD4: In which country were you born? (1= United States of America; 2= Mexico; 3= Canada; 4= United Kingdom; 5= other)
  \item SD6: What is the HIGHEST level of school or degree you have completed? (1= Less than High School Degree; 2= High School Diploma or General Education Diploma (GED); 3= Some college courses/Technical degree/ Vocational program/Associate degree; 4= College degree/advanced degree (Bachelors, Masters, Doctorate))
  \item SD7: In the past 30 days, have you used or taken medication for which a prescription is needed?  Include only those products prescribed by a health professional such as a doctor or dentist.  (0= no; 1= yes)
  \item SD8: How many different times did you stay in any hospital overnight or longer during the past 12 months?
  \item SD9: During the past 30 days, for about how many days did poor physical or mental health keep you from doing your usual activities such as self-care, work, or recreation?}
}
\source{
Data for perceived cognitive functioning from the PROsetta Stone Wave 2
dataset.
}
\usage{
cognition
}
\description{
cognition - polytomous example data
data measured with the FACT-cog (subscale perceived cognitive functioning)
sample size: N=1009
}
\references{
Cella D (2017). PROsetta Stone Wave 2. Harvard Dataverse.
DOI: 10.7910/DVN/WZVYZI.
}
\keyword{datasets}
