% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRm_dataprep_functions.R
\name{dataprep}
\alias{dataprep}
\title{This function makes non-exported functions from the package eRm
(datprep_RM(), datprep_PCM(), datprep_RSM() available. The code is
identical to eRm version 1.0.4 and has only been
combined within this function using if-clauses for the respective model
types. Authors of the code are Patrick Mair, Thomas Rusch, Reinhold
Hatzinger, Marco J. Maier & Rudolf Debelak.}
\usage{
dataprep(X, W, sum0, modelType)
}
\arguments{
\item{X}{data matrix with response categories to be converted into
0/1 matrix}

\item{W}{Design matrix}

\item{sum0}{If TRUE, the parameters are normed to sum-0 by specifying an
appropriate W. If FALSE, the first parameter is restricted to 0.}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}
}
\value{
a list containig the values for X, X01,mt_vek W)
}
\description{
This function makes non-exported functions from the package eRm
(datprep_RM(), datprep_PCM(), datprep_RSM() available. The code is
identical to eRm version 1.0.4 and has only been
combined within this function using if-clauses for the respective model
types. Authors of the code are Patrick Mair, Thomas Rusch, Reinhold
Hatzinger, Marco J. Maier & Rudolf Debelak.
}
\keyword{internal}
