% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRm_datacheck_functions.R
\name{datcheck.LRtest}
\alias{datcheck.LRtest}
\title{Sanity checks for LRtest. This function makes the non-exported function
LRtest() from the package eRm available. The code is identical to eRm
version 1.0.4 and has only been combined within this function. Authors of
the code are Patrick Mair, Thomas Rusch, Reinhold Hatzinger,
Marco J. Maier & Rudolf Debelak.}
\usage{
datcheck.LRtest(x, X, model)
}
\arguments{
\item{x}{submatrix (splitted with "splitcr" and called within Xlist)}

\item{X}{original data matrix (from model fit)}

\item{model}{model of type RM, PCM or RSM}
}
\value{
vector with items to be eliminated
}
\description{
Sanity checks for LRtest. This function makes the non-exported function
LRtest() from the package eRm available. The code is identical to eRm
version 1.0.4 and has only been combined within this function. Authors of
the code are Patrick Mair, Thomas Rusch, Reinhold Hatzinger,
Marco J. Maier & Rudolf Debelak.
}
\keyword{internal}
