% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRm_datacheck_functions.R
\name{datcheck}
\alias{datcheck}
\title{This function makes non-exported functions from the package eRm
(prettyPaste(), component.dist(), reachability(), geodist(), symmetrize(),
available. The code is identical to eRm version 1.0.4 and has only been
combined within this function. Authors of the code are Patrick Mair,
Thomas Rusch, Reinhold Hatzinger, Marco J. Maier & Rudolf Debelak.}
\usage{
datcheck(X, W, mpoints, groupvec, model)
}
\arguments{
\item{X}{data matrix with response categories}

\item{W}{design matrix}

\item{mpoints}{number of meaurement points for repeated measures. Not
important for exhaustiveRasch, because only RM, PCM and RSM models
are supported.}

\item{groupvec}{vector defining groups for group contrasts. Not
important for exhaustiveRasch, because only RM, PCM and RSM models
are supported.}

\item{model}{an object of class RM, PCM or RSM.}
}
\value{
a list containing a dfata matrix without excluded items and a
 value for groupcev.
}
\description{
This function makes non-exported functions from the package eRm
(prettyPaste(), component.dist(), reachability(), geodist(), symmetrize(),
available. The code is identical to eRm version 1.0.4 and has only been
combined within this function. Authors of the code are Patrick Mair,
Thomas Rusch, Reinhold Hatzinger, Marco J. Maier & Rudolf Debelak.
}
\keyword{internal}
