% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_control.R
\name{estimation_control}
\alias{estimation_control}
\title{options for fit_rasch()}
\usage{
estimation_control(est = "psychotools", se = TRUE, sum0 = FALSE)
}
\arguments{
\item{est}{a character value defining the estimation function to use.
Possible values: 'psychotools', 'eRm'.}

\item{se}{a boolean value defining if hessian values and standard errors
for the item parameter estimates are to be computed.}

\item{sum0}{a boolean value defining if item parameters are to be
computed in a way that their sum is 0. This parameter is only meaninful, if
the 'est' parameter is 'eRm". Estimates computed using 'psychotools' are
always computed by setting the first parameter 0.}
}
\value{
a list containing the options
}
\description{
options for fit_rasch()
}
