% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_functions.R
\name{expscore}
\alias{expscore}
\title{returns a matrix with dims like resp matrix with expected scores and more ...
func. by joerg-henrik heine jhheine(at)googlemail.com
needs func. \code{pvx} in i.pvx.R and \code{pvx.matrix} in i.pvx.matrix.R
Notation and formulas see:  Wright & Masters 1982 p.100
in a revised form  (korrigendum) see http://www.rasch.org/rmt/rmt34e.htm}
\usage{
expscore(pers_obj, na_treat = NA)
}
\arguments{
\item{pers_obj}{an object of class "pers"}

\item{na_treat}{internal}
}
\value{
internal
}
\description{
returns a matrix with dims like resp matrix with expected scores and more ...
func. by joerg-henrik heine jhheine(at)googlemail.com
needs func. \code{pvx} in i.pvx.R and \code{pvx.matrix} in i.pvx.matrix.R
Notation and formulas see:  Wright & Masters 1982 p.100
in a revised form  (korrigendum) see http://www.rasch.org/rmt/rmt34e.htm
}
\keyword{internal}
