% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rasch.R
\name{fit_rasch}
\alias{fit_rasch}
\title{parameter estimation for for rasch models.}
\usage{
fit_rasch(X, modelType, estimation_param)
}
\arguments{
\item{X}{a numeric vector containing the index numbers of the items
in dset that are used to fit the model}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}

\item{est}{a character value defining the estimation function to use
(psychtools or eRm for using the estimation function of the respective
package).}
}
\value{
if est=eRm was used: an object of the respecting classes RM, PCM
or RSM of the eRm package (fit rasch models);
if est=psychotools was used: a reduced list of model parameters with the
same names as they are used in the respective eRm object (X, betapar,
etapar, loglik, se.beta, X01, model), but estimated using the respective
function of psychotools.
}
\description{
parameter estimation for for rasch models.
}
\keyword{internal}
