% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit_control.R
\name{itemfit_control}
\alias{itemfit_control}
\title{options for test_itemfit()}
\usage{
itemfit_control(
  zstd = F,
  msq = T,
  outfits = T,
  use.pval = T,
  lowerMSQ = 0.7,
  upperMSQ = 1.3,
  lowerZ = -2,
  upperZ = 2
)
}
\arguments{
\item{zstd}{a boolean value whether to check the z-standardised fit indices}

\item{msq}{aboolean value whether to check the mean-squared fit indices}

\item{outfits}{a boolean value whether to check outfit indices (in FALSE, only infits are checked)}

\item{use.pval}{a boolean value whether to exclude patterns with at least one item with significant p-value}

\item{lowerMSQ}{a numeric value for the lower bound for acceptable Infit (mean-squared fit indices)}

\item{upperMSQ}{a numeric value for the upper bound for acceptable Infit (mean-squared fit indices)}

\item{lowerZ}{a numeric value for the lower bound for acceptable Infit (z-standardised fit indices)}

\item{upperZ}{a numeric value for the upper bound for acceptable Infit (z-standardised fit indices)}
}
\value{
a list containing the options
}
\description{
options for test_itemfit()
}
