% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRm_Mloef_LRtest_Waldtest.R
\name{LRtest}
\alias{LRtest}
\title{Performs Andersen LR-test.
This code is identical to the function of the eRM package (version 0.1.4)
of the same name. Authors of the code are Patrick Mair, Thomas Rusch,
Reinhold Hatzinger, Marco J. Maier & Rudolf Debelak.
The only difference is that it calls the fit_rasch() function for model
fitting instead of the respective functions of eRM.}
\usage{
LRtest(object, splitcr = "median", se = TRUE, estimation_param)
}
\arguments{
\item{object}{object of class RM, PCM or RSM (package eRm)}

\item{splitcr}{splitting criterion for LR-groups. "all.r" corresponds to a complete}

\item{se}{whether standard errors should be computed}

\item{estimation_param}{a list of arguments for estimation control as
returned by the estimation_control() function.}
}
\value{
a list containing the same parameters the respective function of
package eRm would return.
}
\description{
Performs Andersen LR-test.
This code is identical to the function of the eRM package (version 0.1.4)
of the same name. Authors of the code are Patrick Mair, Thomas Rusch,
Reinhold Hatzinger, Marco J. Maier & Rudolf Debelak.
The only difference is that it calls the fit_rasch() function for model
fitting instead of the respective functions of eRM.
}
\keyword{internal}
