% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRm_Mloef_LRtest_Waldtest.R
\name{Mloef}
\alias{Mloef}
\title{This code is identical to the function of the eRM package (version 0.1.4)
of the same name. Authors of the code are Patrick Mair, Thomas Rusch,
Reinhold Hatzinger, Marco J. Maier & Rudolf Debelak.
The only difference is that it calls the fit_rasch() function for model
fitting instead of the respective functions of eRM. Missing values are
not allowed.}
\usage{
Mloef(robj, splitcr = "median", estimation_param)
}
\arguments{
\item{robj}{object of class RM, PCM or RSM (package eRm)}

\item{splitcr}{a character value that defines the splitting criterion for
two groups. "median" (default) and "mean" split items in two groups
according to the median/mean or item raw scores. A vector of length k
(number of items) containing two different elements signifying group
membership of items can be supplied.}

\item{estimation_param}{a list of arguments for estimation constrol as
returned by the estimation_control() function.}
}
\value{
a list containing the same parameters the respective function of
package eRm would return.
}
\description{
This code is identical to the function of the eRM package (version 0.1.4)
of the same name. Authors of the code are Patrick Mair, Thomas Rusch,
Reinhold Hatzinger, Marco J. Maier & Rudolf Debelak.
The only difference is that it calls the fit_rasch() function for model
fitting instead of the respective functions of eRM. Missing values are
not allowed.
}
\keyword{internal}
