% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_control.R
\name{pairwise_control}
\alias{pairwise_control}
\title{options for fitting pairwise models}
\usage{
pairwise_control(resp.cat = NULL, use.thurst = T)
}
\arguments{
\item{resp.cat}{number of response (answer) categories for all items. If not
given, they will be calculated from the data, assuming that every response
category is at least once present in the data.}

\item{use.thurst}{a boolean value defining whether thurstonian threshold
parameters (TRUE, default) or Rasch-Andrich thresholds (step parameters)
will be computed.}
}
\value{
a list containing the options
}
\description{
options for fitting pairwise models
}
