% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passed_exRa.R
\docType{class}
\name{passed_exRa-class}
\alias{passed_exRa-class}
\title{class passed_exRa, an S4 class representing
 an a result of the exhaustive_tests function}
\description{
class passed_exRa, an S4 class representing
 an a result of the exhaustive_tests function
}
\section{Slots}{

\describe{
\item{\code{process}}{a data.frame containg process information
from the call to exhaustive_tests}

\item{\code{passed_combos}}{a list of vectors containing item combinations
using the indices of the items}

\item{\code{passed_models}}{a list of objects of the classes RM, PCM or RSM
of the eRm package (fit rasch models)}

\item{\code{IC}}{a data.frame conatining information criteria for each of the
models in passed_models}

\item{\code{data}}{a data.frame containing the data used for the analyes.}

\item{\code{timings}}{a data.frame containing the the timings of the analyses.}
}}

