% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_psychotools.R
\name{pvx.matrix}
\alias{pvx.matrix}
\title{pvx matrix}
\usage{
pvx.matrix(theta_v, thres, xm_v = NULL)
}
\arguments{
\item{theta_v}{internal}

\item{thres}{internal}

\item{xm_v}{internal}
}
\value{
internal
}
\description{
pvx matrix
}
\keyword{internal}
