% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_psychotools.R
\name{pvx}
\alias{pvx}
\title{this is an internal function for expscore()}
\usage{
pvx(theta, thres, xm = NULL)
}
\arguments{
\item{thres}{internal}

\item{xm}{internal}
}
\value{
internal
}
\description{
this is an internal function for expscore()
}
\keyword{internal}
