% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_functions.R
\name{pvx.super}
\alias{pvx.super}
\title{internal pvx.super function"}
\usage{
pvx.super(theta_v, thres = NULL, dat = NULL)
}
\arguments{
\item{theta_v}{internal}

\item{thres}{internal}

\item{dat}{internal}
}
\value{
internal
}
\description{
internal
}
\keyword{internal}
