% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_passed_xlsx.R
\name{save_passed_xlsx}
\alias{save_passed_xlsx}
\title{saves an Excel file with the results of exhaustive_tests()}
\usage{
save_passed_xlsx(resultData, filename, itemNames = NULL)
}
\arguments{
\item{resultData}{the list object with the result of \link{exhaustive_tests}}

\item{filename}{path and filename of the Excel file to be saved}

\item{itemNames}{the item labels to be used in the Excel sheet, e.g. the columnNames of the data.set that was tested with exhaustive_tests()}
}
\value{
no value is returned. This function saves an Excel file containing 3 sheets: 1) The process information of exhaustive_tests(); 2) a tabulated sheet with all patterns that passed the test circuit of exhaustive_tests(); 3) the information criteria (AIC, BIC, cAIC) of the models that were fit using the respective item patterns.
}
\description{
saves an Excel file with the results of exhaustive_tests()
}
