% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.passed_exRa.R
\name{summary,passed_exRa-method}
\alias{summary,passed_exRa-method}
\title{summary function for class passed_exRa}
\usage{
\S4method{summary}{passed_exRa}(object)
}
\arguments{
\item{object}{an object of class passed_exRa}
}
\description{
summary function for class passed_exRa
}
