% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_itemfit.R
\name{test_itemfit}
\alias{test_itemfit}
\title{checks the itemfit indices of a rasch model using the itemfit() function of eRm.}
\usage{
test_itemfit(
  werte = NULL,
  dset = NULL,
  lower = 0.7,
  upper = 1.3,
  model = NULL,
  p.val = TRUE,
  modelType = NULL
)
}
\arguments{
\item{werte}{a numeric vector containing the index numbers of the items in dset that are used to fit the model}

\item{dset}{a data.frame containing the data}

\item{lower}{lower bound for acceptable Infit (MSQ fit indices are used).}

\item{upper}{upper bound for acceptable Infit (MSQ fit indices are used).}

\item{model}{a list of type RM, PCM or RSM (a previously fit model) matching the value of modelType. If model is provided, this model ist used. If NULL, a model is fit using dset and werte.}

\item{p.val}{a boolean value indicating whether to exclude patterns with at least one item with significant p-value (p<0.05)}

\item{modelType}{a character value defining the rasch model to fit. Possible values: RM, PCM, RSM}
}
\value{
if all fit indices meet the given criteria, a list containing two elements is returned: the pattern that was tested an a list of type RM, RCM or RSM (depending on modelType) with the fit model. If at least one item's fit indices do not meet the given criteria, NULL is returned.
}
\description{
checks the itemfit indices of a rasch model using the itemfit() function of eRm.
}
