% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_pca.R
\name{test_pca}
\alias{test_pca}
\title{runs a principal component analysis (PCA) for binary data based on a tetrachoric correlation matrix}
\usage{
test_pca(werte = NULL, dset = NULL)
}
\arguments{
\item{werte}{a numeric vector containing the index numbers of the items in dset that are used to fit the model}

\item{dset}{a data.frame containing the data}
}
\value{
if all items load on one
}
\description{
runs a principal component analysis (PCA) for binary data based on a tetrachoric correlation matrix
}
