% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_personsItems.R
\name{test_personsItems}
\alias{test_personsItems}
\title{checks the relationship between the person parameter distribution and the item (or: threshold) locations for defined criteria}
\usage{
test_personsItems(
  items = NULL,
  dset = NULL,
  na.rm = T,
  model = NULL,
  modelType = NULL,
  gap_prop = 0,
  extremes = T
)
}
\arguments{
\item{items}{a numeric vector containing the index numbers of the items in dset that are used to fit the model}

\item{dset}{a data.frame containing the data}

\item{na.rm}{a boolean value. If TRUE, all cases with any NA are removed (na.omit). If FALSE, only cases with full NA responses are removed}

\item{model}{a list of type RM, PCM or RSM (a previously fit model) matching the value of modelType. If model is provided, this model ist used. If NULL, a model is fit using dset and items.}

\item{modelType}{a character value defining the rasch model to fit. Possible values: RM, PCM, RSM}

\item{gap_prop}{a numeric value between 0 and 1 that sets the criterion for the minimum proportion of neighboring person parameters with an item/threshold location in between. If set to 0, this criterion will not be checked.}

\item{extremes}{a boolean value indicating if a check for the item/threshold locations left of the 2nd lowest and right of the 2nd highest person parameter}
}
\value{
if the criteria are met, a list containing two elements is returned: the pattern that was tested an a list of type RM, RCM or RSM (depending on modelType) with the fit model. If the criteria are not met, NULL is returned.
}
\description{
checks the relationship between the person parameter distribution and the item (or: threshold) locations for defined criteria
}
