% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_PSI.R
\name{test_PSI}
\alias{test_PSI}
\title{runs a principal component analysis (PCA) on the residuals of the
 rasch model.}
\usage{
test_PSI(
  items = NULL,
  dset = NULL,
  na.rm = TRUE,
  model = NULL,
  p.par = NULL,
  modelType = NULL,
  PSI = NULL,
  estimation_param = NULL,
  pair_param = NULL
)
}
\arguments{
\item{items}{a numeric vector containing the index numbers of the items
in dset that are used to fit the model}

\item{dset}{a data.frame containing the data}

\item{na.rm}{a boolean value. If TRUE, all cases with any NA are removed
(na.omit). If FALSE, only cases with full NA responses are removed}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}

\item{PSI}{a numeric value defining the minimum value for the person-
separation-index (separation reliablility).}

\item{estimation_param}{options for parameter estimation using
\link{estimation_control}}

\item{pair_param}{options for options for fitting pairwise models using
\link{pairwise_control}}
}
\value{
if the maximum eigenvalue of the contrasts of the pca
 is < max_contrast a   #'  list containing two elements is returned:
 the item combination that was tested and a list of type RM, RCM or RSM
 (depending on modelType) with the fit model. Else, NULL is returned.
}
\description{
runs a principal component analysis (PCA) on the residuals of the
 rasch model.
}
\keyword{internal}
