% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_ICs.R
\name{add_ICs}
\alias{add_ICs}
\title{adds information criteria to the @IC slot of an object of
 class \link{passed_exRa-class}}
\usage{
add_ICs(obj, ignoreCores = 1)
}
\arguments{
\item{obj}{an object of class \link{passed_exRa-class}}

\item{ignoreCores}{a numeric value for the number of cpu cores to hold out
in parallelizing computing of the information criteria.}
}
\description{
adds information criteria to the @IC slot of an object of
 class \link{passed_exRa-class}
}
