% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_rawscores.R
\name{all_rawscores}
\alias{all_rawscores}
\title{checks if all possible raw scores occur in the data.}
\usage{
all_rawscores(
  items = NULL,
  dset = NULL,
  na.rm = TRUE,
  model = NULL,
  p.par = NULL,
  modelType = NULL,
  estimation_param = NULL
)
}
\arguments{
\item{items}{a numeric vector containing the index numbers of the items in
dset that are used to fit the model}

\item{dset}{a data.frame containing the data}

\item{na.rm}{a boolean value. If TRUE, all cases with any NA are removed
(na.omit). If FALSE, only cases with full NA responses are removed}

\item{model}{a list of type RM, PCM or RSM (a previously fit model)
matching the value of modelType. If model is provided, this model is used.
 If NULL, a model is fit using dset and items.}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}

\item{estimation_param}{options for parameter estimation using
\link{estimation_control}}
}
\value{
if all possible raw scores occur in dset, a list containing two
 elements is returned: the pattern that was tested an a list of type
  RM, RCM or RSM (depending on modelType) with the fit model. If at least
   one raw score does not occur in dset, NULL is returned.
}
\description{
checks if all possible raw scores occur in the data.
}
\keyword{internal}
