% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation_control.R
\name{estimation_control}
\alias{estimation_control}
\title{options for item parameter estimation}
\usage{
estimation_control(est = "psychotools", resp.cat = NULL, use.thurst = T)
}
\arguments{
\item{est}{a character value defining the estimation function to use.
Possible values: 'psychotools', 'eRm'.}

\item{resp.cat}{number of response (answer) categories for all items.
If not given, they will be calculated from the data, assuming that
every response category is at least once present in the data. Currently
only used for pairwise estimations (m parameter in pairwise::pair).}

\item{use.thurst}{a boolean value defining whether thurstonian threshold
parameters (TRUE, default) or Rasch-Andrich thresholds (step parameters)
will be computed.}
}
\value{
a list containing the options
}
\description{
options for item parameter estimation
}
