% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_rasch.R
\name{fit_rasch}
\alias{fit_rasch}
\title{parameter estimation for for rasch models.}
\usage{
fit_rasch(X, modelType, estimation_param)
}
\arguments{
\item{X}{a numeric vector containing the index numbers of the items
in dset that are used to fit the model}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}

\item{estimation_param}{options for parameter estimation using
\link{estimation_control}}
}
\value{
a model of the class depending of modelType and
estimation_param$est.If no model can be fit, NULL is returened.
}
\description{
parameter estimation for for rasch models.
}
\keyword{internal}
