% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itemfit_control.R
\name{itemfit_control}
\alias{itemfit_control}
\title{options for test_itemfit()}
\usage{
itemfit_control(
  zstd = FALSE,
  msq = TRUE,
  outfits = TRUE,
  use.pval = TRUE,
  lowerMSQ = 0.7,
  upperMSQ = 1.3,
  lowerZ = -1.96,
  upperZ = 1.96,
  use.rel = FALSE
)
}
\arguments{
\item{zstd}{a boolean value whether to check the standardised fit indices}

\item{msq}{a boolean value whether to check the mean-squared fit indices}

\item{outfits}{a boolean value whether to check outfit indices
(if FALSE, only infits are checked)}

\item{use.pval}{a boolean value whether to exclude combinations with at
least one item with significant p-value}

\item{lowerMSQ}{a numeric value for the lower bound for acceptable fit
(mean-squared fit indices)}

\item{upperMSQ}{a numeric value for the upper bound for acceptable fit
(mean-squared fit indices)}

\item{lowerZ}{a numeric value for the lower bound for acceptable fit
(standardised fit indices)}

\item{upperZ}{a numeric value for the upper bound for acceptable fit
(standardised fit indices)}

\item{use.rel}{a boolean value wheter to use the unweighted
(default, FALSE) or weighted item fit indices for MSQ and z-standardised
 fit-indices. Only available for 'pairwise' and 'psychotools' estiomation,
 will be ignored, for 'eRm' estimation.}
}
\value{
a list containing the options
}
\description{
options for test_itemfit()
}
