% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_psychotools.R
\name{LRtest.psy}
\alias{LRtest.psy}
\title{itemfit statistics for psychotools
Anderson's likelihood ratio test for psychotools}
\usage{
LRtest.psy(model, modelType, splitcr = "median", splitseed = NULL)
}
\arguments{
\item{model}{a list of type RM, PCM or RSM (a previously fit model)
matching the value of modelType. If model is provided, this model is used.
 If NULL, a model is fit using dset and items.}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}

\item{splitcr}{as defined by eRm::MLoef: Split criterion to define the
item groups. "median" and "mean" split items in two groups based on their
 items' raw scores. splitcr can also be a vector of length k (where k
 denotes the number of items) that takes two or more distinct values to
  define groups used for the Martin-Löf Test.
A random split, as in pairwise, is also a possible option.}

\item{splitseed}{seed for random split}
}
\value{
a list containing a) a list of item combinations that passed the
 actual test; and b) a list containing the fit models
  of type RM, PCM or RSM.
}
\description{
itemfit statistics for psychotools
Anderson's likelihood ratio test for psychotools
}
\keyword{internal}
