% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_psychotools.R
\name{ppar.psy}
\alias{ppar.psy}
\title{itemfit statistics for psychotools}
\usage{
ppar.psy(model = NULL)
}
\arguments{
\item{model}{an object of 'psychotools' class 'raschmodel', 'pcmodel' or
'rsmodel'  (a model previously fit using the 'psychotools' package)
 matching the value of modelType.}
}
\value{
an object containing person parameters, residuals and PSI
}
\description{
itemfit statistics for psychotools
}
\examples{
 model <- psychotools::raschmodel(ADL[c(6,7,12,14,15)])
 ppar.psy(model)
}
