% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.passed_exRa.R
\name{summary,passed_exRa-method}
\alias{summary,passed_exRa-method}
\title{summary function for class passed_exRa}
\usage{
\S4method{summary}{passed_exRa}(object)
}
\arguments{
\item{object}{an object of \link{passed_exRa-class}}
}
\value{
no object is returned, the summary is printed to the console. This
comprises: tested scale lengths, total number of tested item combinations,
applied tests, number of passed item combinations that passed each test, item
importance and runtimes.
}
\description{
summary function for class passed_exRa
}
