% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mloef.R
\name{test_mloef}
\alias{test_mloef}
\title{runs Martin-Loef Test with median split using the MLoef() function of eRm.}
\usage{
test_mloef(werte = NULL, dset = NULL, model = NULL, modelType = NULL)
}
\arguments{
\item{werte}{a numeric vector containing the index numbers of the items in dset that are used to fit the model}

\item{dset}{a data.frame containing the data}

\item{model}{a list of type RM, PCM or RSM (a previously fit model) matching the value of modelType. If model is provided, this model ist used. If NULL, a model is fit using dset and werte.}

\item{modelType}{a character value defining the rasch model to fit. Possible values: RM, PCM, RSM}
}
\value{
if the p-value of the test is not significant (above p=0.05), a list containing two elements is returned: the pattern that was tested an a list of type RM, RCM or RSM (depending on modelType) with the fit model. If the test is significant (p<0.05), NULL is returned.
}
\description{
runs Martin-Loef Test with median split using the MLoef() function of eRm.
}
