% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_psychotools.R
\name{waldtest.psy}
\alias{waldtest.psy}
\title{Fischer and Scheiblechner's "wald-like" S-statistic for psychotools}
\usage{
waldtest.psy(
  model,
  modelType,
  splitcr = "median",
  splitseed = NULL,
  icat = FALSE
)
}
\arguments{
\item{model}{a list of type RM, PCM or RSM (a previously fit model)
matching the value of modelType. If model is provided, this model is used.
 If NULL, a model is fit using dset and items.}

\item{modelType}{a character value defining the rasch model to fit.
Possible values: RM, PCM, RSM}

\item{splitcr}{as defined by eRm::Waldtest: Split criterion for subject
raw score splitting. median uses the median as split criterion, mean
 performs a mean-split. Optionally splitcr can also be a dichotomous
  vector which assigns each person to a certain subgroup
   (e.g., following an external criterion). This vector can be numeric,
    character or a factor.
A random split, as in pairwise, is also a possible option.}

\item{splitseed}{seed for random split}

\item{icat}{a boolean value defining wheter to use item parameters
(psychotools function itempar, if TRUE) or item category parameters (psychotools
function threshpar)}
}
\value{
a list containing a) a list of item combinations that passed the
 actual test; and b) a list containing the fit models
  of type RM, PCM or RSM.
}
\description{
Fischer and Scheiblechner's "wald-like" S-statistic for psychotools
}
